Add filter to remove offline devices from `DeviceList` and persist filters with `SceneStorage`
